sdk/sdk-python/pyproject.toml
[tool.poetry]
name = "directus"
version = "0.1.0"
description = "Python wrapper for the Directus API"
authors = ["Jo <ljonas@riseup.net>"]
license = "GPL-3.0"
repository = "https://github.com/directus/sdk-python"

[tool.poetry.dependencies]
python = "^3.7"
pyjwt = "^1.7.1"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.782"
pylint = "^2.5.3"
pytest = "^5.4.3"
responses = "^0.10.15"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
